+++ src/components/ConfirmDialog.jsx
@@ -1,5 +1,7 @@
 import "../styles/ConfirmDialog.css";
+import { highlightText } from "../utils/highlightText";
 import FormButton from "./FormButton";
+import { v4 as generateId } from "uuid";
 
 export default function ConfirmDialog({prompt, actionText, onConfirm, onCancel})
 {
@@ -13,7 +15,16 @@ export default function ConfirmDialog({prompt, actionText, onConfirm, onCancel})
             data-testid="backdrop"
         >
             <div role="dialog" className="dialog-box">
-                <p className="prompt">{prompt}</p>
+                <p className="prompt">
+                {
+                    highlightText(prompt, "*").map(segment => {
+
+                      if (segment.highlight) return <span key={generateId()} className="red-highlight">{segment.value}</span>  
+
+                      return segment.value
+                    })
+                }
+                </p>
                 <FormButton
                     classes={["form-button", "white-button"]}
                     onClick={e => {