+++ src/pages/Links.jsx
@@ -78,7 +78,7 @@ export default function Links({linksItems, setLinksItems, setDialogState,  empty
                                 setDialogState({
                                     open: true,
                                     actionText: "Delete",
-                                    prompt: "Are you sure you want to premenantly delete this item from the links section?",
+                                    prompt: "Are you sure you want to *premenantly delete* this item from the links section?",
                                     onConfirm: () => {
                                         setLinksItems(deleteItem(id, linksItems))
                                     }